Entrega variaveis de comunicação do projeto com Firebase
