+++ src/hooks/useFetchDocument.jsx
@@ -1,0 +1,8 @@
+import { useState, useEffect } from "react"
+import {db} from "../firebase/config"
+import {doc, getDoc} from "firebase/firestore"
+
+export const useFetchDocument = () =>{
+
+    
+}
