+++ src/hooks/useFetchDocuments.jsx
@@ -1,0 +1,8 @@
+import { useState, useEffect } from "react"
+import {db} from "../firebase/config"
+import { collet, query, ordyBy, onSnapshot, where } from "firebase/firestore"
+
+export const useFetchDocuments = () =>{
+
+    
+}
