+++ src/hooks/useQuery.jsx
@@ -1,0 +1,6 @@
+import {useLocation} from "react-router-dom"
+import { useMemo } from "react"
+
+export function useQuery() {
+    
+}
