+++ src/hooks/useUpdateDocument.jsx
@@ -1,0 +1,8 @@
+import { useState, useEffect, useReducer } from "react"
+import {db} from "../firebase/config"
+import {doc, updateDoc} from "firebase/firestore"
+
+export const useUpdateDocument = () =>{
+
+    
+}
